+++ CHANGELOG.md
@@ -9,9 +9,13 @@
 
 ### Bugfixes
 
-- Do not include non-GET route in the sitemap
-
 </details>
+
+## 0.5.1
+
+### Bugfixes
+
+- Do not include non-GET route in the sitemap
 
 ## 0.5.0